requirements.txt
Flask==2.1.1
uwsgi==2.0.20
PyYAML==6.0
psycopg2==2.9.3